refactor: split VoronoiFrag.ts into sub-modules, clean up stale comments (#201)
